normalizer/requirements.txt
certifi==2019.9.11
chardet==3.0.4
CouchDB==1.2
idna==2.8
paho-mqtt==1.4.0
redis==3.3.8
requests==2.22.0
SQLAlchemy==1.3.9
urllib3==1.25.6
